Remove open redirection exploit
By creating URL like the following ones:

https://ipdata.info/%5Cexample.com/

It is possible to create a link that redirect to a arbitraty page while
still starting as a known website.

This commit addresses this exploit.

It does change the behaviour a bit though, as instead of giving a
permanent redirect to the version without trailing slashes, it just
accepts URL with trailing slashes.